RStoolbox {RStoolbox}	R Documentation

RStoolbox: A Collection of Remote Sensing Tools

Description

The RStoolbox package provides a set of functions which simplify performing standard remote sensing tasks in R.

Data Import and Export

readMeta: import Landsat metadata from MTL or XML files

stackMeta: load Landsat bands based on metadata

readSLI & writeSLI: read and write ENVI spectral libraries

saveRSTBX & readRSTBX: save and re-import RStoolbox classification objects (model and map)

readEE: import and tidy EarthExplorer search results

Data Pre-Processing

radCor: radiometric conversions and corrections. Primarily, yet not exclusively, intended for Landsat data processing. DN to radiance to reflectance conversion as well as DOS approaches

topCor: topographic illumination correction

cloudMask & cloudShadowMask: mask clouds and cloud shadows in Landsat or other imagery which comes with a thermal band

classifyQA: extract layers from Landsat 8 QA bands, e.g. cloud confidence

rescaleImage: rescale image to match min/max from another image or a specified min/max range

normImage: normalize imagery by centering and scaling

histMatch: matches the histograms of two scenes

coregisterImages: co-register images based on mutual information

panSharpen: sharpen a coarse resolution image with a high resolution image (typically panchromatic)

Data Analysis

spectralIndices: calculate a set of predefined multispectral indices like NDVI

tasseledCap: tasseled cap transformation

sam: spectral angle mapper

rasterPCA: principal components transform for raster data

rasterCVA: change vector analysis

unsuperClass: unsupervised classification

superClass: supervised classification

fCover: fractional cover of coarse resolution imagery based on high resolution classification

Data Display

ggR: single raster layer plotting with ggplot2

ggRGB: efficient plotting of remote sensing imagery in RGB with ggplot2

[Package RStoolbox version 0.4.0 Index]