cloudMask { RStoolbox}	R Documentation

Very simple cloud detection for imagery with blue and thermal...

Description

Very simple cloud detection for imagery with blue and thermal bands

Usage

cloudMask(x, threshold, minCloudSize, windowSize1 = 5, windowSize2 = 11,
  maskGrowing = TRUE, sanitize = TRUE, lowBand = "B1", tirBand = "B6",
  plot = TRUE, verbose = TRUE, returnDiffLayer = FALSE)

Arguments

x
RasterBrick or RasterStack with reflectance and brightness temperature OR the mask of a previous run of cloudMask with returnDiffLayer=TRUE.

threshold
cloud detection threshold. If not provided it will be guessed. Everything *above* this threshold will be considered a cloud pixel (unless it is removed by filtering afterwards).

minCloudSize
minimum number of cloud pixels in window1

windowSize1
odd number, rectangular moving window to remove clouds which are too small (likely artefacts)

windowSize2
odd number, rectangular buffer around cluster centers

maskGrowing
logical. Applies simple region-growing (rectangular buffering) to the cloud mask. If TRUE windowSize2 must be specified.

sanitize
logical. Should small clouds (possibly false positives) be removed by filtering? If TRUE windowSize1 must be specified.

lowBand
bandname or number for the blue band

tirBand
bandname or number for the thermal band

plot
logical. Provides plots of the cloud mask for all sub-steps (sanitizing etc.) Helpful to find proper parametrization.

verbose
logical. Print messages or suppress.

returnDiffLayer
logical. If TRUE, the difference layer will be returned along with the cloudmask. This option allows to re-use the difference layer in cloudMask.

Note

Typically clouds are cold in the thermal region and have high reflectance in short wavelengths (blue). By differencing the two bands and thresholding a rough cloud mask can be obtained. More sophisticated approaches can be found elsewhere, e.g. fmask.

It can make sense to find a suitable threshold on a cropped version of the scene. Also make sure you make use of the returnDiffLayer argument to save yourself one processing step. Sanitizing and region growing can be seen as final polishing, i.e. as long as the pure cloud centers are not detected properly, you can turn those two arguments off if they take too long to calculate. Once your mask detects obvious cloud pixels properly re-enable sanitizing and regionGrowing for fine tuning if desired. Finally, once a suitable threshold is established re-run cloudMask on the whole scene with this threshold and go get a coffee.

Examples

ls <- stackMeta("path/to/MTL.txt")

#> Warning in file(file, "rt"): cannot open file 'path/to/MTL.txt': No such
#> file or directory

#> Error in file(file, "rt"): cannot open the connection

ls_cor <- radCor(ls, "path/to/MTL.txt")

#> Warning in file(file, "rt"): cannot open file 'path/to/MTL.txt': No such
#> file or directory

#> Error in file(file, "rt"): cannot open the connection

ls_cmask <-cloudMask(ls_cor, returnDiffLayer = TRUE)

#> Error in match(x, table, nomatch = 0L): object 'ls_cor' not found

[Package RStoolbox version 0.0.0.9000 Index]