fCover { RStoolbox}	R Documentation

Estimate fractional cover using image regression

Description

fCover takes a classified high resolution image, e.g. vegetation and non-vegetation based on Landsat and calculates cover fractions for pixels of a coarser resolution, e.g. MODIS.

Usage

fCover(classImage, predImage, nSamples = 1000, classes = 1, model = "rf",
  tuneLength = 3, tuneGrid = NULL, method = "cv", maxNA = 0.1,
  filename = NULL, verbose, ...)

Arguments

classImage
high resolution RasterLayer containing a landcover classification, e.g. as obtained by superClass.

predImage
coearse resolution RasterLayer for which fractional cover will be estimated.

nSamples
Integer. Number of pixels to sample from predImage to train the regression model

classes
Integer. Classes for which fractional cover should be estimated.

model
Character. Which model to fit for image regression. See train for options. Defaults to randomForest ('rf')

tuneLength
Integer. Number of levels for each tuning parameters that should be generated by train.

tuneGrid
Optional data.frame with exact tuning values. The columns are named the same as the tuning parameters.

method
Character. Resampling method for parameter tuning. Defaults to 10 fold cross-validation. See trainControl for options.

maxNA
Numeric. Maximal proportion of NAs allowed in training pixels.

filename
Character. Raster filename.

verbose
Logical. Print progress information.

...
further arguments to be passed to trainControl and writeRaster

Details

fCover gets the pixel values in a high resolution classified image that correspond to individual randomly selected moderate resolution pixels and then calculates the percent of the classified image pixels that represent your cover type of interest. In other words, if your high resolution image has a pixel size of 1m and your moderate resolution image has a pixel size of 30m the sampling process would take a block of 900 of the 1m resolution pixels that correspond to a single 30m pixel and calculate the percentage of the 1m pixels that are forest. For example, if there were 600 forest pixels and 300 non-forest pixels the value given for the output pixel would be 0.67 since 67

fCover relies on train from the caret package which provides acces to a huge number of classifiers. Please see the available options at train. The default classifier (randomForest) we chose has been shown to provide very good results in image regression and hence its recomended you start with this one. If you choose a different one, make sure it can run in regression mode.

If you register a parallel backend before, model fitting will run in parallel.

If the total no-data values for a block of high resolution pixels is greater than maxNA then it will not be included in the training data set since there is too much missing data to provide a reliable cover percentage. If the no-data proporton is less then maxNA the no-data pixels are removed from the total number of pixels when calculating the percent cover.

Value

Returns a list with two elements: models contains the fitted models evaluated in tenfold cross-validation (caret train objects); fCover contains a RasterStack with a fractional cover layer for each requested class

See Also

superClass

Examples

## Not run: 
##D ## Create fake input images
##D lsat <- brick(system.file("external/rlogo.grd", package="raster"))
##D agg.level <- 9
##D modis <- aggregate(lsat, agg.level)
##D 
##D ## Perform classification
##D lc    <- unsuperClass(lsat, nClass=2)
##D 
##D ## Calculate the true cover (possible in this example because the fake corse resolution imagery
##D ## is exactly res(lsat)*9
##D trueCover <- aggregate(lc$map, agg.level, fun = function(x, ...){sum(x == 1, ...)/sum(!is.na(x))})
##D 
##D ## Run with randomForest and support vector machine (radial basis kernel)
##D ## Of course the SVM is handicapped in this example due to poor tuning (tuneLength)
##D par(mfrow=c(2,3))
##D for(model in c("rf", "svmRadial")){
##D    fc <- fCover(
##D            classImage = lc$map ,
##D            predImage = modis,
##D            classes=1,
##D            model=model,
##D            nSample = 50,
##D            number = 5,
##D            tuneLength=2
##D    )
##D 
##D    ## How close is it to the truth?
##D    compare.rf <- trueCover - fc$map
##D    plot(fc$map, main = paste("Fractional Cover: Class 1\nModel:", model))
##D    plot(compare.rf, main = "Diffence\n true vs. predicted")
##D    plot(trueCover[],fc$map[],  xlim = c(0,1), ylim =c(0,1),
##D            xlab = "True Cover", ylab = "Predicted Cover" )
##D    abline(coef=c(0,1))
##D    rmse <- sqrt(cellStats(compare.rf^2, sum))/ncell(compare.rf)
##D    r2 <- cor(trueCover[], fc$map[], "complete.obs")
##D    text(0.9,0.1,paste0(paste(c("RMSE:","R2:"),
##D                            round(c(rmse, r2),3)),collapse="\n"), adj=1)
##D }
##D 
##D ## Reset par
##D par(mfrow=c(1,1))
## End(Not run)

[Package RStoolbox version 0.0.0.9000 Index]