getValidation {RStoolbox}	R Documentation

Extract validation results from superClass objects

Description

Extract validation results from superClass objects

Usage

getValidation(x, from = "testset", metrics = "overall")

Arguments

x
superClass object or caret::confusionMatrix

from
Character. 'testset' extracts the results from independent validation with testset. 'cv' extracts cross-validation results.

metrics
Character. Only relevant in classification mode (ignored for regression models). Select 'overall' for overall accuracy metrics, 'classwise' for classwise metrics, 'confmat' for the confusion matrix itself and 'caret' to return the whole caret::confusionMatrix object.

Value

Returns a data.frame with validation results. If metrics = 'confmat' or 'caret' will return a table or the full caret::confusionMatrix object, respectively.

Examples

library(pls)
## Fit classifier (splitting training into 70% training data, 30% validation data)
train <- readRDS(system.file("external/trainingPoints.rds", package="RStoolbox"))
SC   <- superClass(rlogo, trainData = train, responseCol = "class",
                    model="pls", trainPartition = 0.7)

#> 09:17:00 | Begin sampling training data

#> 09:17:00 | Starting to fit model

#> 09:17:00 | Starting spatial predict

#> 09:17:00 | Begin validation

#> ******************** Model summary ********************

#> Partial Least Squares 
#> 
#> 21 samples
#>  3 predictor
#>  3 classes: 'A', 'B', 'C' 
#> 
#> No pre-processing
#> Resampling: Cross-Validated (5 fold) 
#> Summary of sample sizes: 16, 17, 16, 17, 18 
#> Resampling results across tuning parameters:
#> 
#>   ncomp  Accuracy   Kappa    
#>   1      0.6733333  0.5000000
#>   2      0.8600000  0.7633987
#> 
#> Accuracy was used to select the optimal model using the largest value.
#> The final value used for the model was ncomp = 2.
#> [[1]]
#>   TrainAccuracy TrainKappa method
#> 1          0.86  0.7633987    pls
#> 
#> [[2]]
#> Cross-Validated (5 fold) Confusion Matrix 
#> 
#> (entries are average cell counts across resamples)
#>  
#>           Reference
#> Prediction   A   B   C
#>          A 1.4 0.6 0.0
#>          B 0.0 0.8 0.0
#>          C 0.0 0.0 1.4
#>                             
#>  Accuracy (average) : 0.8571

#> ******************** Validation summary ********************

#> Confusion Matrix and Statistics
#> 
#>           Reference
#> Prediction A B C
#>          A 3 1 0
#>          B 0 2 0
#>          C 0 0 3
#> 
#> Overall Statistics
#>                                           
#>                Accuracy : 0.8889          
#>                  95% CI : (0.5175, 0.9972)
#>     No Information Rate : 0.3333          
#>     P-Value [Acc > NIR] : 0.0009653       
#>                                           
#>                   Kappa : 0.8333          
#>                                           
#>  Mcnemar's Test P-Value : NA              
#> 
#> Statistics by Class:
#> 
#>                      Class: A Class: B Class: C
#> Sensitivity            1.0000   0.6667   1.0000
#> Specificity            0.8333   1.0000   1.0000
#> Pos Pred Value         0.7500   1.0000   1.0000
#> Neg Pred Value         1.0000   0.8571   1.0000
#> Prevalence             0.3333   0.3333   0.3333
#> Detection Rate         0.3333   0.2222   0.3333
#> Detection Prevalence   0.4444   0.2222   0.3333
#> Balanced Accuracy      0.9167   0.8333   1.0000

## Independent testset-validation
getValidation(SC)

#>   model validation  Accuracy     Kappa AccuracyLower AccuracyUpper AccuracyNull AccuracyPValue McnemarPValue
#> 1   pls    testset 0.8888889 0.8333333     0.5175035     0.9971909    0.3333333      0.0009653           NaN

getValidation(SC, metrics = "classwise")

#>   model validation class Sensitivity Specificity Pos.Pred.Value Neg.Pred.Value Precision    Recall        F1
#> 1   pls    testset     A   1.0000000   0.8333333           0.75      1.0000000      0.75 1.0000000 0.8571429
#> 2   pls    testset     B   0.6666667   1.0000000           1.00      0.8571429      1.00 0.6666667 0.8000000
#> 3   pls    testset     C   1.0000000   1.0000000           1.00      1.0000000      1.00 1.0000000 1.0000000
#>   Prevalence Detection.Rate Detection.Prevalence Balanced.Accuracy
#> 1  0.3333333      0.3333333            0.4444444         0.9166667
#> 2  0.3333333      0.2222222            0.2222222         0.8333333
#> 3  0.3333333      0.3333333            0.3333333         1.0000000

## Cross-validation based 
getValidation(SC, from = "cv")

#>   model validation  Accuracy     Kappa AccuracyLower AccuracyUpper AccuracyNull AccuracyPValue McnemarPValue
#> 1   pls         cv 0.8571429 0.7857143      0.636576      0.969511    0.3333333   1.101588e-06           NaN

[Package RStoolbox version 0.4.0 Index]