rsOpts {RStoolbox}	R Documentation

Set global options for RStoolbox

Description

shortcut to options(RStoolbox.*)

Usage

rsOpts(verbose)

Arguments

verbose
Logical. If TRUE many functions will print status messages about the current processing step. By default verbose mode is disabled.

Value

No return, just a setter for the verbosiness of the RStoolbox package

Examples

rsOpts(verbose=TRUE)

[Package RStoolbox version 0.4.0 Index]