sam {RStoolbox}	R Documentation

Spectral Angle Mapper

Description

Calculates the angle in spectral space between pixels and a set of reference spectra (endmembers) for image classification based on spectral similarity.

Usage

sam(img, em, angles = FALSE, ...)

Arguments

img
RasterBrick or RasterStack or SpatRaster. Remote sensing imagery.

em
Matrix or data.frame with endmembers. Each row should contain the endmember spectrum of a class, i.e. columns correspond to bands in img. It is reccomended to set the rownames to class names.

angles
Logical. If TRUE a RasterBrick containing each one layer per endmember will be returned containing the spectral angles.

...
further arguments to be passed to writeRaster

Details

For each pixel the spectral angle mapper calculates the angle between the vector defined by the pixel values and each endmember vector. The result of this is one raster layer for each endmember containing the spectral angle. The smaller the spectral angle the more similar a pixel is to a given endmember class. In a second step one can the go ahead an enforce thresholds of maximum angles or simply classify each pixel to the most similar endmember.

Value

SpatRaster If angles = FALSE a single Layer will be returned in which each pixel is assigned to the closest endmember class (integer pixel values correspond to row order of em.

Examples

library(terra)
library(ggplot2)

## Sample endmember spectra
## First location is water, second is open agricultural vegetation
pts <- data.frame(x = c(624720, 627480), y = c(-414690, -411090))
endmembers <- extract(lsat, pts)
rownames(endmembers) <- c("water", "vegetation")

#> Error in `rownames<-`(`*tmp*`, value = c("water", "vegetation")): attempt to set 'rownames' on an object with no dimensions

## Calculate spectral angles
lsat_sam <- sam(lsat, endmembers, angles = TRUE)

#> Error in sam(lsat, endmembers, angles = TRUE): The number of columns in em must match the number of bands in x.

plot(lsat_sam)

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'lsat_sam' not found

## Classify based on minimum angle
lsat_sam <- sam(lsat, endmembers, angles = FALSE)

#> Error in sam(lsat, endmembers, angles = FALSE): The number of columns in em must match the number of bands in x.

ggR(lsat_sam, forceCat = TRUE, geom_raster=TRUE) +
             scale_fill_manual(values = c("blue", "green"), labels = c("water", "vegetation"))

#> Error in .toTerra(img): object 'lsat_sam' not found

[Package RStoolbox version 0.4.0 Index]