topCor { RStoolbox}	R Documentation

Topographic Illumination Correction

Description

Topographic Illumination Correction

Usage

topCor(img, dem, metaData, solarAngles, method = "C", stratImg,
  stratMethod = "noStrat", nStrat = 5, illu)

Arguments

img
Raster*. Imagery to correct

dem
Raster*. Either a digital elevation model as a RasterLayer or a RasterStack/Brick with pre-calculated slope and aspect (see terrain) in which case the layers must be named 'slope' and 'aspect'.

metaData
Character, ImageMetaData. Either a path to a Landsat meta-data file (MTL) or an ImageMetaData object (see readMeta)

solarAngles
Numeric vector containing sun azimuth and sun zenith (in radians and in that order). Not needed if metaData is provided

method
Character. One of c("cos", "avgcos", "minnaert", "C", "stat", "illu"). Choosing 'illu' will return only the local illumination map.

stratImg
RasterLayer by which to stratify by, e.g. NDVI. Defaults to 'slope'

stratMethod
Character. One of c("noStrat", "stratEqualBins", "stratQuantiles").

nStrat
Integer. Number of bins or quantiles to stratify by. If a bin has less than 50 samples it will be merged with the next bin.

illu
Raster*. Optional pre-calculated ilumination map. Run topCor with method="illu" to calculate one.

[Package RStoolbox version 0.0.0.9000 Index]