radCor { RStoolbox}	R Documentation

Radiometric calibration and correction

Description

Implements several different methods for absolute radiometric correction of Landsat data. You can either specify a metadata file, or supply all neccesary values manually. With proper parametrization APREF and SDOS should work for other sensors as well.

Usage

radCor(img, metaData, radiance = FALSE, method = "APREF",
  bandSet = "full", SHV, hazeBand, atHaze, darkProp = 0.02, verbose)

Arguments

img
raster object

metaData
object of class ImageMetaData or a path to the meta data (MTL) file.

radiance
Logical. If TRUE only radiance will be calculated. Otherwise reflectance and brightness temperature are calculated.

method
Radiometric correction method to be used. There are currently four methods available (see Details): "APREF", "SDOS", "DOS", "COSTZ".

bandSet
numeric or character. original Landsat band numbers or names in the form of ("B1", "B2" etc). If set to 'full' all bands in the solar region will be processed.

SHV
starting haze value, can be estimated using estimateSHV(). if not provided and method is "DOS" or "COSTZ" SHV will be estimated in an automated fashion. Not needed for apparent reflectance.

hazeBand
band from which SHV was estimated.

atHaze
character. Scene haze characteristics. Will be estimated if not expicilty provided. Must be one of "veryClear", "clear", "moderate", "hazy" or "veryHazy".

darkProp
numeric. Estimated proportion of dark pixels in the scene. Used only for automatic guessing of SHV.

verbose
Logical. Print status information.

Details

Conversion to top of atmosphere radiance (W/(m^2 * srad * μ m)) Conversion to at-satellite brightness temperature (K) Conversion to top of atmosphere reflectance (unitless) corrected for the sun angle Estimation of at-surface spectral reflectance (unitless)

APREF

Apparent reflectance

DOS

Dark object subtratction following Chavez (1989)

COSTZ

Dark object subtraction following Chaves(1996)

SDOS

Simple dark object subtraction. Classical DOS, Lhaze must be estimated for each band separately.

The implemented sun-earth distances neglect the earth's eccentricity. Instead it uses a 100 year daily average (1979-2070).

Atmospheric haze decay model according to Chavez (1989)

veryClear

λ^{-4.0}

clear

λ^{-2.0}

moderate

λ^{-1.0}

hazy

λ^{-0.7}

veryHazy

λ^{-0.5}

Note

This was originally a fork of randcorr in the landsat package. It may be slower, however it works on Raster* objects and hence is memory-safe.

References

S. Goslee (2011): Analyzing Remote Sensing Data in R: The landsat Package. Journal of Statistical Software 43(4).

[Package RStoolbox version 0.0.0.9000 Index]