readMeta { RStoolbox}	R Documentation

Read landsat MTL metadata files

Description

Besides reading metadata, readMeta deals with legacy versions of Landsat metadata files and where possible adds missing information (radiometric gain and offset, earth-sun distance).

Usage

readMeta(file, raw = FALSE)

Arguments

file
path to Landsat MTL file (...MTL.txt)

raw
Logical. If TRUE the full raw metadate will be returned as a list. All important metadata are homogenized into a standard format (ImageMetaData) and some information is added.

Value

Object of class ImageMetaData

[Package RStoolbox version 0.0.0.9000 Index]