saveRSTBX { RStoolbox}	R Documentation

Save and Read RStoolbox Classification Results

Description

Saves objects of classes unsuperClass, superClass, rasterPCA and fCover to file. Useful to archive the fitted models.

Usage

saveRSTBX(x, filename, format = "raster", ...)

readRSTBX(filename)

Arguments

x
RStoolbox object of classes c("fCover", "rasterPCA", "superClass", "unsuperClass")

filename
Character. Path and filename. Any file extension will be ignored.

format
Character. Driver to use for the raster file

...
further arguments passed to writeRaster

Value

The output of writeRSTBX will be at least two files written to disk: a) an .rds file containing the object itself and b) the raster file (depending on the driver you choose this can be more than two files).

Functions

readRSTBX:

Note

All files must be kept in the same directory to read the full object back into R by means of readRSTBX. You can move them to another location but you'll have to move *all* of them (just like you would with Shapefiles). In case the raster file(s) is missing, readRSTBX will still return the object but the raster will be missing.

writeRSTBX and readRSTBX are convenience wrappers around saveRDS, readRDS. This means you can read all files created this way also with base functionality as long as you don't move your files. This is because x$map is a Raster* object and hence contains only a static link to the file on disk.

Examples

input <- brick(system.file("external/rlogo.grd", package="raster"))
## Create filename
file  <- paste0(tempdir(), "/test", runif(1))
## Run PCA
rpc   <- rasterPCA(input, filename = file, nSample = 100)

#> Error in .startRasterWriting(x, filename, format = filetype, ...): /tmp/Rtmp1cOoYQ/test0.456437604967505.grd exists. use 'overwrite=TRUE' if you want to overwrite it

## Save object
saveRSTBX(rpc, filename=file)

#> Error in .startRasterWriting(x, filename, format = filetype, ...): /tmp/Rtmp1cOoYQ/test0.456437604967505.grd exists. use 'overwrite=TRUE' if you want to overwrite it

## Which files were written?
list.files(tempdir(), pattern = basename(file))

#> [1] "test0.456437604967505.grd" "test0.456437604967505.gri"
#> [3] "test0.456437604967505.rds"

## Re-read files
re_rpc <- readRSTBX(file)
## Compare
all.equal(re_rpc, rpc)

#>  [1] "Component \"call\": target, current do not match when deparsed"                                                                                                
#>  [2] "Component \"model\": Component \"sdev\": Mean relative difference: 0.04471185"                                                                                 
#>  [3] "Component \"model\": Component \"loadings\": Mean relative difference: 0.01170357"                                                                             
#>  [4] "Component \"model\": Component \"center\": Mean relative difference: 0.03463799"                                                                               
#>  [5] "Component \"model\": Component \"n.obs\": Modes: numeric, logical"                                                                                             
#>  [6] "Component \"model\": Component \"n.obs\": target is numeric, current is logical"                                                                               
#>  [7] "Component \"model\": Component \"call\": target, current do not match when deparsed"                                                                           
#>  [8] "Component \"map\": Attributes: < Component \"data\": Attributes: < Component \"fromdisk\": 1 element mismatch > >"                                             
#>  [9] "Component \"map\": Attributes: < Component \"data\": Attributes: < Component \"inmemory\": 1 element mismatch > >"                                             
#> [10] "Component \"map\": Attributes: < Component \"data\": Attributes: < Component \"max\": Mean relative difference: 0.03365062 > >"                                
#> [11] "Component \"map\": Attributes: < Component \"data\": Attributes: < Component \"min\": Mean relative difference: 0.09283748 > >"                                
#> [12] "Component \"map\": Attributes: < Component \"data\": Attributes: < Component \"values\": Modes: logical, numeric > >"                                          
#> [13] "Component \"map\": Attributes: < Component \"data\": Attributes: < Component \"values\": Attributes: < Component \"dim\": Mean absolute difference: 3890 > > >"
#> [14] "Component \"map\": Attributes: < Component \"data\": Attributes: < Component \"values\": Lengths (0, 23331) differ (comparison on first 0 components) > >"     
#> [15] "Component \"map\": Attributes: < Component \"file\": Attributes: < Component \"bandorder\": names for target but not for current > >"                          
#> [16] "Component \"map\": Attributes: < Component \"file\": Attributes: < Component \"byteorder\": names for target but not for current > >"                          
#> [17] "Component \"map\": Attributes: < Component \"file\": Attributes: < Component \"datanotation\": 1 string mismatch > >"                                          
#> [18] "Component \"map\": Attributes: < Component \"file\": Attributes: < Component \"driver\": 1 string mismatch > >"                                                
#> [19] "Component \"map\": Attributes: < Component \"file\": Attributes: < Component \"name\": 1 string mismatch > >"                                                  
#> [20] "Component \"map\": Attributes: < Component \"file\": Attributes: < Component \"nbands\": Mean relative difference: 0.6666667 > >"                              
#> [21] "Component \"map\": Attributes: < Component \"file\": Attributes: < Component \"nodatavalue\": Mean relative difference: Inf > >"

file.remove(file)

#> Warning in file.remove(file): cannot remove file
#> '/tmp/Rtmp1cOoYQ/test0.456437604967505', reason 'No such file or
#> directory'

#> [1] FALSE

[Package RStoolbox version 0.0.0.9000 Index]