stackMeta { RStoolbox}	R Documentation

Import separate Landsat files into single stack

Description

Reads Landsat MTL or XML metadata files and loads single Landsat Tiffs into a rasterStack. Be aware that by default stackLS() does NOT import panchromatic bands nor thermal bands with resolutions != 30m.

Usage

stackMeta(file, allResolutions = FALSE, quantity = "all",
  category = "image")

Arguments

file
character. Path to Landsat MTL metadata file (not an XML file!).

allResolutions
logical. if TRUE a list will be returned with length = unique spatial resolutions.

quantity
Character vector. Which quantity should be returned. Options: digital numbers ('dn'), top of atmosphere reflectance ('tre'), at surface reflectance ('sre'), brightness temperature ('bt'), spectral index ('idx').

category
character vector. Which category of data to return. Options 'image': image data, 'pan': panchromatic image, 'index': multiband indices, 'qa' quality flag bands, 'all': all categories.

Value

Either a list of rasterStacks comprising all resolutions or only one rasterStack comprising only 30m resolution imagery

Note

Be aware that by default stackLS() does NOT import panchromatic bands nor thermal bands with resolutions != 30m. Use the allResolutions argument to import all layers. Note that nowadays the USGS uses cubic convolution to resample the TIR bands to 30m resolution.

[Package RStoolbox version 0.0.0.9000 Index]